admin bugfix and refactoring

+++ app/Http/Controllers/Admin/PostController.php
@@ -93,8 +93,9 @@ class PostController extends Controller
 
         $post = Post::find($id);
         $data = $request->all();
-        $data['thumbnail'] = Post::uploadImage($request, $post->thumbnail);
-
+        if ($file = Post::uploadImage($request, $post->thumbnail)) {
+            $data['thumbnail'] = $file;
+        }
         $post->update($data);
         $post->tags()->sync($request->tags);
         
@@ -111,7 +112,10 @@ class PostController extends Controller
     {
         $post = Post::find($id);
         $post->tags()->sync([]);
-        Storage::delete($post->thumbnail);
+
+        if ($post->thumbnail)
+            Storage::delete($post->thumbnail);
+            
         $post->delete();
         return redirect()->route('posts.index')->with('success', 'Статья удалена');
     }
